+++ Modules/Sale/Database/Migrations/2021_07_31_212446_create_sale_details_table.php
@@ -22,7 +22,7 @@ class CreateSaleDetailsTable extends Migration
             $table->integer('quantity');
             $table->integer('price');
             $table->integer('unit_price');
-            $table->integer('sub_total');
+            $table->unsignedBigInteger('sub_total');
             $table->integer('product_discount_amount');
             $table->string('product_discount_type')->default('fixed');
             $table->integer('product_tax_amount');
