+++ Modules/Sale/Database/Migrations/2021_08_07_192203_create_sale_payments_table.php
@@ -16,7 +16,7 @@ class CreateSalePaymentsTable extends Migration
         Schema::create('sale_payments', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('sale_id');
-            $table->integer('amount');
+            $table->unsignedBigInteger('amount');
             $table->date('date');
             $table->string('reference');
             $table->string('payment_method');
